rag_service/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-dotenv
certifi
motor
pymongo
pydantic>=2.0.0,<3.0.0
python-multipart
google-genai==1.30.0
pypdf
python-jose[cryptography]
numpy
qdrant-client>=1.9.0
rank-bm25
groq>=0.9.0
tenacity>=8.2.0
trafilatura>=2.0.0
